group 2/3

+++ src/i-adapters/controllers/Group_deleteController.ts
@@ -5,5 +5,8 @@ export type IGroupDeleteController = ReturnType<typeof group_deleteController>
 export const group_deleteController = (
         groupService : IGroupRepository
     ) => async (group_id : number) : Promise<boolean> => {
-        return await groupService.deleteGroup(group_id)
+        if (group_id != 1) {
+            return await groupService.deleteGroup(group_id)
+        }
+        return false;
     }
